[ElementTiming] Add request URL and responseEnd
This CL adds the initial URL and the response end from ResourceTiming
into ElementTiming. The information is currently only stored on the
ResourceTimingInfo object which is not accessible to ImageElementTiming
(and in fact, could be gone by the time the ElementTiming code is
executed). Thus, we store the information in ImageResourceContent.

Bug: 879270
Change-Id: I72a80c67956adb6e869f39ca3c77829bad86044e